pyproject.toml
[tool.poetry]
name = "hawk-tui"
version = "0.1.0"
description = "Connect to that thang"
authors = ["David Daniel <davydany@gmail.com>"]
license = "GPL-3.0-or-later"
readme = "README.md"
homepage = "https://github.com/davydany/hawk-tui"
repository = "https://github.com/davydany/hawk-tui"
keywords = ["database", "tui", "connector"]
packages = [
    { include = "hawk_tui", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.12.3"
textual = "^0.27.0"
click = "^8.1.3"
psycopg2-binary = "^2.9.5"
mysql-connector-python = "^8.0.32"
kafka-python = "^2.0.2"
redis = "^4.5.1"
elasticsearch = "^8.6.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
hawk-tui = "hawk_tui.main:hawk"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.flake8]
max-line-length = 88
extend-ignore = "E203"